split validation schema
