+++ deploy/tools/envs-validator/schemas/features/apiDocs.ts
@@ -1,0 +1,17 @@
+import { urlTest } from "../../utils";
+import * as yup from 'yup';
+import { replaceQuotes } from 'configs/app/utils';
+import type { ApiDocsTabId } from 'types/views/apiDocs';
+import { API_DOCS_TABS } from 'types/views/apiDocs';
+
+export const apiDocsSchema = yup
+  .object()
+  .shape({
+    NEXT_PUBLIC_API_DOCS_TABS: yup.array()
+      .transform(replaceQuotes)
+      .json()
+      .of(yup.string<ApiDocsTabId>().oneOf(API_DOCS_TABS)),
+    NEXT_PUBLIC_API_SPEC_URL: yup
+      .string()
+      .test(urlTest),
+  });
