+++ deploy/tools/envs-validator/schemas/features/index.ts
@@ -1,0 +1,12 @@
+export * from './address3rdPartyWidgets';
+export * from './ads';
+export * from './apiDocs';
+export * from './beaconChain';
+export * from './bridgedToken';
+export * from './defiDropdown';
+export * from './marketplace';
+export * from './megaEth';
+export * from './rollup';
+export * from './tac';
+export * from './userOps';
+export * from './zetaChain';
