+++ deploy/tools/envs-validator/test/.env.multichain
@@ -26,3 +26,8 @@ NEXT_PUBLIC_IS_TESTNET=true
 NEXT_PUBLIC_USE_NEXT_JS_PROXY=true
 NEXT_PUBLIC_HAS_USER_OPS=true
 NEXT_PUBLIC_ADVANCED_FILTER_ENABLED=false
+
+NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID=xxx
+NEXT_PUBLIC_GOOGLE_ANALYTICS_PROPERTY_ID=xxx
+NEXT_PUBLIC_ROLLBAR_CLIENT_TOKEN=xxx
+NEXT_PUBLIC_MIXPANEL_PROJECT_TOKEN=xxx
